Rates module

+++ db/migrate/20221216124518_create_user_favourites.rb
@@ -1,8 +1,8 @@
 class CreateUserFavourites < ActiveRecord::Migration[7.0]
   def change
     create_table :user_favourites do |t|
-      t.integer :user_id
-      t.integer :topic_id
+      t.bigint :user_id
+      t.bigint :topic_id
 
       t.timestamps
     end
